refactor(minifier): move dce conditional expression to `RemoveDeadCode` (#5971)
This is aligned to closure compiler
